+++ app/models/hud_reports/question_sheet.rb
@@ -5,12 +5,14 @@
 ###
 
 # utility class to dry up answers
-module HudApr::Generators::Shared::Fy2024
+module HudReports
   class QuestionSheet
     attr_reader :report, :question
 
     def initialize(report:, question:)
       @report = report
+      raise "invalid question id #{question.inpsect}" unless question =~ /\AQ\d+/
+
       @question = question
     end
 
@@ -107,30 +109,6 @@ module HudApr::Generators::Shared::Fy2024
     def next_column
       max = @headers.keys.max
       max ? Util.next_column_identifier(max) : 'A'
-    end
-  end
-
-  # internal class
-  class Util
-    # given a column (letter), return the next col in sequence
-    # A => B, Z => AA, etc
-    def self.next_column_identifier(column)
-      # Convert the column to an integer
-      n = 0
-      column.each_char do |char|
-        n = n * 26 + (char.ord - 65 + 1) # 65 is the ASCII code for 'A', add 1 for 1-based indexing
-      end
-
-      # Increment the integer and convert it back to an identifier
-      n += 1
-      result = ''
-      while n.positive?
-        remainder = (n - 1) % 26 # Subtracting 1 to handle 1-based indexing
-        result = (65 + remainder).chr + result # 65 is the ASCII code for 'A'
-        n = (n - 1) / 26 # Subtracting 1 to handle 1-based indexing
-      end
-
-      result
     end
   end
