+++ drivers/hud_apr/app/models/hud_apr/generators/shared/fy2024/question_twenty_five.rb
@@ -129,7 +129,7 @@ module HudApr::Generators::Shared::Fy2024
 
     private def q25c_veteran_gender
       gender_question(
-        question: '25c',
+        question: 'Q25c',
         members: universe.members.where(veteran_clause),
         populations: q25_populations,
       )
